@@ -9,11 +9,11 @@
   },
   "license": "MIT",
   "dependencies": {
-    "react": "^16.2.0",
+    "react": "^16.4.1",
     "react-ace": "^6.1.0",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.4.1",
     "react-onclickout": "^2.0.8",
-    "react-router-dom": "^4.2.2"
+    "react-router-dom": "^4.3.1"
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
